refactored the jaas client call back handler class, and added utility clazz
